// frontend/agro-sensor-hub-webapp/src/types/sensorActivity.ts
export interface SensorActivity {
    id: number;
    device_id: string;
    zone: string;
    env_humidity: number;
    env_temperature: number;
    ground_sensor_1: number;
    ground_sensor_2: number;
    ground_sensor_3: number;
    ground_sensor_4: number;
    ground_sensor_5: number;
    ground_sensor_6: number;
    created_at: string;
} 